requirements.txt
boto3
pandas
psycopg2
SQLAlchemy
pyarrow
gdown
python-dotenv
